+++ qlcc/src/main/java/com/mbs/qlcc/adapters/request/Apartment/UpdateApartmentRequest.java
@@ -1,6 +1,8 @@
 package com.mbs.qlcc.adapters.request.Apartment;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -8,8 +10,8 @@ import lombok.*;
 public class UpdateApartmentRequest {
     private int floor;
     private String aptNumber;
-    private Double grossArea;
-    private Double coefficient;
+    private BigDecimal grossArea;
+    private BigDecimal coefficient;
     private String aptType;
     private String description;
 }
