+++ qlcc/src/main/java/com/mbs/qlcc/entities/Apartment/IApartmentFactory.java
@@ -1,7 +1,9 @@
 package com.mbs.qlcc.entities.Apartment;
 
+import java.math.BigDecimal;
+
 public interface IApartmentFactory {
     Apartment create(String buildingId, String complexId, int floor,
-                     String aptNumber, Double grossArea, Double coefficient,
+                     String aptNumber, BigDecimal grossArea, BigDecimal coefficient,
                      String aptType, String description);
 }
