+++ qlcc/src/main/java/com/mbs/qlcc/usecases/request/Apartment/UpdateApartmentInpRequest.java
@@ -1,17 +1,20 @@
 package com.mbs.qlcc.usecases.request.Apartment;
+
+import java.math.BigDecimal;
+
 public class UpdateApartmentInpRequest {
     private int floor;
     private String aptNumber;
-    private Double grossArea;
-    private Double coefficient;
+    private BigDecimal grossArea;
+    private BigDecimal coefficient;
     private String aptType;
     private String description;
 
     public UpdateApartmentInpRequest() {
     }
 
-    public UpdateApartmentInpRequest(int floor, String aptNumber, Double grossArea,
-                                     Double coefficient, String aptType, String description) {
+    public UpdateApartmentInpRequest(int floor, String aptNumber, BigDecimal grossArea,
+                                     BigDecimal coefficient, String aptType, String description) {
         this.floor = floor;
         this.aptNumber = aptNumber;
         this.grossArea = grossArea;
@@ -36,19 +39,19 @@ public class UpdateApartmentInpRequest {
         this.aptNumber = aptNumber;
     }
 
-    public Double getGrossArea() {
+    public BigDecimal getGrossArea() {
         return grossArea;
     }
 
-    public void setGrossArea(Double grossArea) {
+    public void setGrossArea(BigDecimal grossArea) {
         this.grossArea = grossArea;
     }
 
-    public Double getCoefficient() {
+    public BigDecimal getCoefficient() {
         return coefficient;
     }
 
-    public void setCoefficient(Double coefficient) {
+    public void setCoefficient(BigDecimal coefficient) {
         this.coefficient = coefficient;
     }
